move bottom fade to the hooks file
